feat: build voting app
*Description*

Why:

Wanted a simple voting app that allowed for write in votes, limited number of candidates, and had a time limit for voting.
Also wanted to validate if there's an existing candidate with a similar name given a write in vote and, if so, assign a vote for that candidate.

This change addresses the need by:

* Uses Levenshtein distance algorithm to check distance between two names.
* Caps number of candidates
* Only allows a given user to vote once
* Refreshes vote screen every 10 seconds

Uses a simple postgres database with two tables to track both votes and voters (but we are not tracking who voted for who).